+++ rEFIt_UEFI/Platform/b64cdecode.cpp
@@ -41,7 +41,7 @@ long base64_decode_block(const char* code_in, const int length_in, char* plainte
 				{
 					state_in->step = step_a;
 					state_in->plainchar = *plainchar;
-					return plainchar - plaintext_out;
+					return (long)(plainchar - plaintext_out); // we assume that plainchar - plaintext_out cannot be > MAX_LONG
 				}
 				fragment = base64_decode_value(*codechar++);
 			} while (fragment < 0);
@@ -52,7 +52,7 @@ long base64_decode_block(const char* code_in, const int length_in, char* plainte
 				{
 					state_in->step = step_b;
 					state_in->plainchar = *plainchar;
-					return plainchar - plaintext_out;
+					return (long)(plainchar - plaintext_out); // we assume that plainchar - plaintext_out cannot be > MAX_LONG
 				}
 				fragment = base64_decode_value(*codechar++);
 			} while (fragment < 0);
@@ -64,7 +64,7 @@ long base64_decode_block(const char* code_in, const int length_in, char* plainte
 				{
 					state_in->step = step_c;
 					state_in->plainchar = *plainchar;
-					return plainchar - plaintext_out;
+					return (long)(plainchar - plaintext_out); // we assume that plainchar - plaintext_out cannot be > MAX_LONG
 				}
 				fragment = base64_decode_value(*codechar++);
 			} while (fragment < 0);
@@ -76,7 +76,7 @@ long base64_decode_block(const char* code_in, const int length_in, char* plainte
 				{
 					state_in->step = step_d;
 					state_in->plainchar = *plainchar;
-					return plainchar - plaintext_out;
+					return (long)(plainchar - plaintext_out); // we assume that plainchar - plaintext_out cannot be > MAX_LONG
 				}
 				fragment = base64_decode_value(*codechar++);
 			} while (fragment < 0);
@@ -84,7 +84,7 @@ long base64_decode_block(const char* code_in, const int length_in, char* plainte
 		}
 	}
 	/* control should not reach here */
-	return plainchar - plaintext_out;
+	return (long)(plainchar - plaintext_out); // we assume that plainchar - plaintext_out cannot be > MAX_LONG
 }
